Put recipe links on the recipe home page in a list
Put all of the relative links to each recipe page in an unordered
list on the Odin recipe home page. Arranging the relative links in a
list format is more visually appealing compared to having them next to
each other horizontally. It is also easier for page visitors to
differentiate the links from one another.